+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * First release.
 * List users assigned to a given role. Refs UIROLES-31.
 * Retrieve up to 5000 capabilities at once. Refs UIROLES-47.
+* Include `settings.authorization-roles.enabled` permission set. Refs UIROLES-54.
+