+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change history for ui-authorization-roles
 
+## 1.3.0 IN PROGRESS
+
+* List users assigned to a given role. Refs UIROLES-31.
+* Retrieve up to 5000 capabilities at once. Refs UIROLES-47.
+* Include `settings.authorization-roles.enabled` permission set. Refs UIROLES-54. 
+
 ## 1.2.0
 
 * First release.
-* List users assigned to a given role. Refs UIROLES-31.
-* Retrieve up to 5000 capabilities at once. Refs UIROLES-47.
-* Include `settings.authorization-roles.enabled` permission set. Refs UIROLES-54.